update: neovim and plugins, fixed a couple things...
So updating broke lspconfig. Rather than figuring out the native
lsp setups right now, we just manually reverted using our old
lazylock version.

- conjure also supports js now which is cool, but seems to be
experimental as we would get a boatload of errors upon opening
a js file, but the repl interaction seems to work.
Still we only want to use it for lisp so we added the filetypes
table in our conjure.config. Well that didn't work, it's an after
plugin and we're assuming it would then be too late to say what
filetypes should be supported (and their autocmds setup), so we
looked around and found that you can specify an init function
in the conjure config directly. That works.

Transparency broke for our floating windows. Thought to look
at the catppucin repo and there it was!

Everything else is fine. :)